backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.2
django-cors-headers==4.5.0
django-bootstrap3==24.2
django-rest-knox==5.0.2
djangorestframework-simplejwt==5.3.1
dj-database-url==2.2.0
gunicorn==22.0.0
whitenoise==6.8.2
python-dotenv==1.0.1
psycopg2-binary==2.9.10
Pillow
python-dotenv